cafeteria_connect/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Deprecated==1.2.18
Django==5.2.3
django-debug-toolbar==5.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
img2pdf==0.6.1
kafka-python==2.2.15
kombu==5.5.4
lxml==6.0.0
packaging==25.0
pikepdf==9.9.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==6.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
